+++ dal/models/translation.py
@@ -1,0 +1,9 @@
+"""Translation model"""
+from .model import Model
+
+
+class Translation(Model):
+    """Translation model"""
+
+
+Model.register_model_class("Translation", Translation)