+++ CHANGELOG.md
@@ -1,10 +1,13 @@
-# v3.4.2
+# vTBD
 - [BP-1456](https://movai.atlassian.net/browse/BP-1456): Add support for user language configuration
   - Introduced `language` attribute for `BaseUser` and `InternalUser` models.
   - Added validation for `language` using `VALID_LANGUAGES` from `movai_core_shared.consts`.
   - Updated JSON schemas (`InternalUser.json` and `User.json`) to include the `Language` field.
   - Added getter and setter for the `language` property in `BaseUser`.
  
+# vTBD
+- [BP-1463](https://movai.atlassian.net/browse/BP-1463): Create translation scope
+
 # v3.4.1
 - Fix mobdata usage (missing 1 required positional argument: 'args')